project/fall_detection/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tkinter-tooltip>=2.0.0
Pillow>=10.0.0
scipy>=1.11.0
imutils>=0.5.4
psutil>=5.9.0
smtplib
email
threading
time 